server/requirements.txt
flask
flask-cors
firebase_admin
python-dotenv